Split TIS analysis into multiple files
