+++ openpathsampling/analysis/__init__.py
@@ -1,2 +1,3 @@
 from .path_histogram import PathHistogram
+from . import tis
 from . import tools
