+++ openpathsampling/analysis/tis/tis_analysis.py
@@ -1,0 +1,18 @@
+import collections
+import openpathsampling as paths
+from openpathsampling.netcdfplus import StorableNamedObject
+from openpathsampling.numerics import LookupFunction
+import pandas as pd
+import numpy as np
+
+from .core import *
+
+########## GENERAL HISTOGRAMMING
+
+
+
+
+############### HISTOGRAMMING MAX LAMBDA
+
+
+
